+++ R/mkin_wide_to_long.R
@@ -1,4 +1,4 @@
-if(getRversion() >= '2.15.1') utils::globalVariables(c("name", "time", "value"))
+utils::globalVariables(c("name", "time", "value"))
 
 #' Convert a dataframe with observations over time into long format
 #'